Update anagolay-project-idiyanale-phase-1.md (#1010)
* Update anagolay-project-idiyanale-phase-1.md

As per @cruikshankss recommendation and advice, i am updating the grant with the new expense we had throughout the milestones. I didn't touch M1 since it is closed and paid out, so i put the difference in the M2.  I updated the actual time that it took us to finish this grant.

* reduce the cost from 15 to 7.5